% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{gnis_get}
\alias{gnis_get}
\title{gnis_get}
\usage{
gnis_get(fname = "", state = "", county = "", cell = "", ftype = "")
}
\arguments{
\item{fname}{character}

\item{state}{character}

\item{county}{character}

\item{cell}{character}

\item{ftype}{character}
}
\description{
gnis_get
}
\examples{
\dontrun{
gnis_get(fname = "Lake Michigan", state = "Michigan", ftype = "lake")
gnis_get(fname = "Big Lake", state = "Arizona", ftype = "lake")
}
}

